+++ packages/kysely/src/codes/get.ts
@@ -5,19 +5,22 @@ import { Code, CodeType } from "@authhero/adapter-interfaces";
 
 export function get(db: Kysely<Database>) {
   return async (
-    _: string,
+    tenant_id: string,
     code_id: string,
     code_type: CodeType,
   ): Promise<Code | null> => {
-    const code = await db
+    let query = db
       .selectFrom("codes")
-      // We currently don't have the tenant_id in all places
-      // .where("codes.tenant_id", "=", tenant_id)
       .where("codes.code_id", "=", code_id)
       .where("codes.code_type", "=", code_type)
-      .where("codes.expires_at", ">", new Date().toISOString())
-      .selectAll()
-      .executeTakeFirst();
+      .where("codes.expires_at", ">", new Date().toISOString());
+
+    // The tenant_id isn't always avaialble when using the same domain
+    if (tenant_id.length) {
+      query = query.where("codes.tenant_id", "=", tenant_id);
+    }
+
+    const code = await query.selectAll().executeTakeFirst();
 
     if (!code) {
       return null;